Move fertilize and breed actions to drake reducer
Part of work to clean up the root reducer to make further refactoring
easier.

Both these actions only affect the `drakes` property, so can be handled
entirely by the drakes reducer.

The one issue is that `fertilize` needs additional information from
elsewhere in the state tree, namely the gametes. While we could maybe
completely restructure the state, or add an all-purpose way for all
reducers to get the entire state (https://github.com/reactjs/redux/issues/601#issuecomment-174387073)
it seemed more natural simply to let the drakes reducer know about the
gametes. This pattern can be applied elsewhere when a reducer is only
updating one property, but needs knowledge about other state props.
