+++ src/code/reducers/index.js
@@ -9,7 +9,7 @@ import routing from './routing';
 import moves from './moves';
 import modalDialog from './modal-dialog';
 import userDrakeHidden from './user-drake-hidden';
-import gametes, { motherCurrentGamete, fatherCurrentGamete } from '../modules/gametes';
+import gametes from '../modules/gametes';
 import drakes from './drakes';
 import baskets from './baskets';
 import notifications from '../modules/notifications';
@@ -31,45 +31,6 @@ function initialState() {
           });
 }
 
-function fertilize(state) {
-  let chromosomes0 = new BioLogica.Organism(BioLogica.Species.Drake, state.drakes[0].alleleString, state.drakes[0].sex).getGenotype().chromosomes,
-      chromosomes1 = new BioLogica.Organism(BioLogica.Species.Drake, state.drakes[1].alleleString, state.drakes[1].sex).getGenotype().chromosomes,
-      alleleString = "",
-      sex = 1,
-      side, chromosome, name;
-  for (name in chromosomes0) {
-    side = motherCurrentGamete(state.gametes)[name];
-    chromosome = chromosomes0[name][side];
-    if (chromosome && chromosome.alleles) alleleString += "a:" + chromosome.alleles.join(",a:") + ",";
-  }
-  for (name in chromosomes1) {
-    side = fatherCurrentGamete(state.gametes)[name];
-    if (side === "y") sex = 0;
-    chromosome = chromosomes1[name][side];
-    if (chromosome && chromosome.alleles && chromosome.alleles.length) alleleString += "b:" + chromosome.alleles.join(",b:") + ",";
-  }
-
-  return state.setIn(["drakes", 2], {
-    alleleString,
-    sex
-  });
-}
-
-function breedClutch(state, clutchSize) {
-  let mother = new BioLogica.Organism(BioLogica.Species.Drake, state.drakes[0].alleleString, state.drakes[0].sex),
-      father = new BioLogica.Organism(BioLogica.Species.Drake, state.drakes[1].alleleString, state.drakes[1].sex),
-      newState = state;
-
-  for (let i = 0; i < clutchSize; i++) {
-    let clutchDrake = BioLogica.breed(mother, father, true);
-    newState = newState.setIn(["drakes", newState.drakes.length], {
-      alleleString: clutchDrake.getAlleleString(),
-      sex: clutchDrake.sex
-    });
-  }
-  return newState;
-}
-
 export default function reducer(state, action) {
   if (!state) state = initialState();
 
@@ -79,7 +40,7 @@ export default function reducer(state, action) {
     modalDialog: modalDialog(state.modalDialog, action),
     userDrakeHidden: userDrakeHidden(state.userDrakeHidden, action),
     gametes: gametes(state.gametes, action),
-    drakes: drakes(state.drakes, action),
+    drakes: drakes(state.drakes, state.gametes, action),
     baskets: baskets(state.baskets, action),
     notifications: notifications(state.notifications, action)
   });
@@ -94,11 +55,6 @@ export default function reducer(state, action) {
         challengeProgress: progress
       });
     }
-    case actionTypes.FERTILIZED:
-      return fertilize(state);
-
-    case actionTypes.CLUTCH_BRED:
-      return breedClutch(state, action.clutchSize);
 
     case actionTypes.OFFSPRING_KEPT: {
       let progress = updateProgress(state, action.success);
